Mostly figured out the AI
Still needs thorough testing but it's good for the most part
Currently there is a small bug with player targeting being overwritten, will investigate later

The Volwraths do not crash anymore and do in fact have turns. This likely affected other enemies as well but it was only known to affect them
The enemy with the least health is the target for healing skills. This will also apply to single target skills that don't heal but apply a state, this is fine for now but should be tweaked in the future.
Skills are removed from the options once selected (if the enemy can take more than one action) so that skills are never repeated during the same turn.
So-called "bad targets" are filtered out when picking who to attack to help prevent wasted turns.
I also prevent the Death state on skills being considered for the Preventative mindset as it should be Aggressive.
A smattering of other small changes across the board