Finalising Militaur and lots of code changes
All of Militaur's skills are in, though they are not thoroughly tested due to specific requirements.
Fixed enemy ally targeting to actually function, by extension also fixed player targets being redirected.
Fixed buffs and debuffs not being properly applied to enemies (how did we never notice this).
Some small optimisations by abstracting away certain calculations. Applies to actor PDR, MDR, and MRG.
Added a popup for when the Prismatic Morphguard increases a resistance and what to, AND when it's capped to 50% for that element.
That's all I can think of, maybe a couple other small things here and there, I had been rooting around in the code for a while.