Lots and lots more code improvements
Fixed Sacred Burial from using mismatched flags
Fixed Hollow Knell causing a Float HP value. It now also refreshes the actor during battle to update the bars straight away. Also changed the HP and MP gain to be based on the MHP and MMP of the enemy that died rather than the holder's.
Fixed Prismatic Morphguard's processing of normal attack so as to ignore the "Physical" element. Also abstracted the resistances to the item_element_rate method so that the Battle_Engine popups correctly show when the element is resisted. Also gave it functionality to gain 2.5% resistance to secondary elements if a skill has one.
Fixed buffs incorrectly adding as debuffs to enemies.
Added ActionMemory to enemies. With it, they remember their actions used over the previous 3 turns. Skills that appear a lot within those turns will get a reduced weight to increase variety.
Added foundation for an enemy to focus a target, like enrage in DQ. This currently only functions to target the ally that is worst afflicted but can be expanded in the future.
Added more things for Mister Bones to give you in return for Artefact Shards (I did this yesterday but forgot to say).
Added Earth/Nature skill. Took way longer than it should have to get it working, but I figured it out. It's neat.
Divine Burst can no longer crit. Sorry.